python/requirements.txt
donna25519
pycryptodome>=3.6.6